+++ src/Vortice.DirectX/DataBox.cs
@@ -1,4 +1,3 @@
-
 // Copyright (c) 2010-2014 SharpDX - Alexandre Mutel
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy