+++ src/Vortice.DirectX/Interop.cs
@@ -10,22 +10,32 @@ namespace Vortice
 {
     public static unsafe class Interop
     {
+        public static IntPtr Read<T>(IntPtr source, ref T value) where T : unmanaged
+        {
+            fixed (void* valuePtr = &value)
+            {
+                Unsafe.CopyBlockUnaligned(valuePtr, (void*)source, (uint)(sizeof(T)));
+                return source + sizeof(T);
+            }
+        }
+
         public static void Read<T>(IntPtr source, T[] values) where T : unmanaged
         {
-            var count = values.Length;
+            int count = values.Length;
             fixed (void* dstPtr = values)
             {
                 Unsafe.CopyBlockUnaligned(dstPtr, (void*)source, (uint)(count * sizeof(T)));
             }
         }
 
-        public static void Write<T>(IntPtr destination, ref T value) where T : unmanaged
+        public static IntPtr Write<T>(IntPtr destination, ref T value) where T : unmanaged
         {
             unsafe
             {
                 fixed (void* valuePtr = &value)
                 {
                     Unsafe.CopyBlockUnaligned((void*)destination, valuePtr, (uint)(sizeof(T)));
+                    return destination + sizeof(T);
                 }
             }
         }
@@ -33,6 +43,11 @@ namespace Vortice
         public static void Write<T>(IntPtr destination, T[] values) where T : unmanaged
         {
             MemoryHelpers.Write(destination, new Span<T>(values), values.Length);
+        }
+
+        public static IntPtr Write<T>(IntPtr destination, T[] values, int offset, int count) where T : unmanaged
+        {
+            return MemoryHelpers.Write(destination, new Span<T>(values).Slice(offset), count);
         }
 
         public static IntPtr Alloc<T>(int count = 1) where T : unmanaged
@@ -50,7 +65,7 @@ namespace Vortice
 
             int structSize = sizeof(T);
             int totalSize = values.Length * structSize;
-            var ptr = Marshal.AllocHGlobal(totalSize);
+            IntPtr ptr = Marshal.AllocHGlobal(totalSize);
 
             var walk = (byte*)ptr;
             for (int i = 0; i < values.Length; i++)